task_manager/requirements.txt
asgiref==3.8.1
Django==5.2.1
<<<<<<< Updated upstream
=======
djangorestframework==3.16.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
>>>>>>> Stashed changes
sqlparse==0.5.3
